Consider tools provided by Visual Studio.
It is frustrating that Visual Studio comes with perfectly good versions of some of our dependencies, but we stubbornly don't use them because they aren't available on a PATH or anything like that.

This change fixes that by looking at the VS instances from vswhere and considering tools therein if they are present.

This change needs approval from the teams that own Git, CMake, and Node in VS before it can land.